requirements.txt
Django==1.8.3
django-bootstrap-toolkit==2.15.0
django-registration-redux==1.2
phileo==1.3
Pillow==2.9.0
sorl-thumbnail==12.3
wheel==0.24.0
